Minor code cleanup: warning

+++ common/src/famfs_lf_connect.c
@@ -55,7 +55,7 @@ int lf_client_init(LF_CL_t *lf_node, N_PARAMS_t *params)
     const char		*pname;
     int			i, rc;
 
-    node = (params->fam_map)? lf_node->fam_id : lf_node->node_id;
+    node = (params->fam_map)? (int)lf_node->fam_id : lf_node->node_id;
     if (params->lf_fabric) {
 	pname = params->lf_fabric;
     } else {
